Bump monorepo dependency to latest `protocol-23` branch (#362)
* Bump image version to current futurenet
* Fetch protocol-23 branch of monorepo

It is known and intentional that the Go XDR is different than
the Rust XDR (check-dependencies failing), because the Rust
side (simulation et al.) is not ready yet, but we need RPC to 
not break on the new meta format.

* Stop running P21 tests because checks fail:
The stellar-core binary is 22.x.x but it supports Protocol 23, 
meaning version validation fails because 22 (core version) < 
23 (minProtocolVersionSupported in the monorepo).

This is a temporary pause. Once P23 is ready we will have 
P22 and P23 runs instead of P21 and 22, anyway.